Change Notes: - Improved organistaion of functions for project loader business code.
